refactor: unify float type names (Float=single, Double=double) and add 1.0f literals
Rename the floating-point type system so single precision is `Float`
(.NET float32) and double precision is `Double` (.NET float). Previously
`Float` meant double and `Single` meant single, which was confusing.

Add literal suffix support: `1.0f` is single-precision (`Float`) and `1.0`
(no suffix) is double-precision (`Double`).

This threads through the whole pipeline: TypeId, lexer/tokens, AST/HIR/
ClosedHir literal nodes, builtin conversions (toSingle/toFloat -> toFloat/
toDouble) and operator tables, MIR immediates, codegen, and the atlalib
serialization/display layers. Examples (hello_role, gui_calc) move to
`Double` to preserve their double-precision behavior.

+++ src/Atla.Core/Semantics/Resolve.fs
@@ -152,7 +152,7 @@ module Resolve =
         moduleScope.DeclareType("Bool", TypeId.Bool)
         moduleScope.DeclareType("Int", TypeId.Int)
         moduleScope.DeclareType("Float", TypeId.Float)
-        moduleScope.DeclareType("Single", TypeId.Single)
+        moduleScope.DeclareType("Double", TypeId.Double)
         moduleScope.DeclareType("String", TypeId.String)
         declareSystemType symbolTable moduleScope "System.Int32" |> ignore
